requirements.txt
torch
torchvision
torchaudio
pillow
numpy
tqdm
nltk
gradio
json
matplotlib
pycocotools
